Add print.js and import to index.js
