+++ src/index.js
@@ -1,3 +1,5 @@
+import './style.css';
+import displayData from './print.js'
 window.onload = function () {
 	const title = document.getElementById('title');
 	const description = document.getElementById('description')
@@ -24,35 +26,7 @@ window.onload = function () {
 
 	}
 
-	const displayData = function (todo) {
-		const list = document.getElementById('content');
-		const item = document.querySelector(`[data-key='${todo._id}']`);
-		if (todo.deleted) {
-			item.remove();
-			return;
-		}
 
-
-		const isChecked = todo._checked ? 'done' : '';
-		const node = document.createElement('li');
-		node.setAttribute('class', `todo-item ${isChecked}`);
-		node.setAttribute('data-key', todo._id);
-		node.innerHTML = `
-
-<span>Title:${todo._title}</span><br>
-<span>${todo._description}</span><br>
-<span>${todo._duedate}</span><br>
-<label for = "${todo._id}" class = 'js-tick'>Checkbox</label>
-<input id = "${todo._id}" type = "checkbox" ><br>
-<button class="delete-todo js-delete-todo">Delete</button>
-`;
-		if (item) {
-			list.replaceChild(node, item);
-		}
-		else {
-			list.append(node);
-		}
-	}
 	const toggleDone = function (key) {
 		const index = todoItems.findIndex(item => item._id === Number(key));
 		todoItems[index]._checked = !todoItems[index]._checked;
